backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==10.0.0
drf-extra-fields==3.7.0
